+++ includes/class-blueworx-deck-builder-plugin.php
@@ -22,7 +22,6 @@ final class Blueworx_Deck_Builder_Plugin {
 		Blueworx_Deck_Builder_Roles::register();
 		Blueworx_Deck_Builder_Types::register();
 		Blueworx_Deck_Builder_Editor::register();
-		Blueworx_Deck_Builder_Library::register();
 		Blueworx_Deck_Builder_Link::register();
 		Blueworx_Deck_Builder_Admin::instance()->register();
 	}